+++ src/layouts/Main.tsx
@@ -18,10 +18,10 @@ import CategoryIcon from '@mui/icons-material/Category';
 import IconButton from '@mui/material/IconButton';
 
 
-import Navigation from 'src/layouts/Navigation/SidebarLeft';
-import Body  from 'src/layouts/Body/Body';
-import SidebarRight from 'src/layouts/SidebarRight/SidebarRight';
-import AudioPlayer from 'src/components/AudioPlayer/AudioPlayer';
+import Navigation from 'src/layouts/components/SidebarLeft/SidebarLeft';
+import Body  from 'src/layouts/components/Body/Body';
+import SidebarRight from 'src/layouts/components/SidebarRight/SidebarRight';
+import AudioPlayer from 'src/layouts/components/AudioPlayer/AudioPlayer';
 
 
 export const drawerWidth = 240;
@@ -45,7 +45,6 @@ export default function Layout({window}: IPropTypes) {
       setMobileOpen(!mobileOpen);
     }
   };
-
 
   return (
     <Box sx={{ display: 'flex', backgroundColor:'#000', color:'#fff' }} >
