--- src/layouts/Main/_test_/Main.test.tsx
@@ -1,3 +1,0 @@
-import {screen, render} from '@testing-library/react';
-
-test.todo('it renders the main')
