+++ src/layouts/components/SidebarLeft/SidebarLeft.tsx
@@ -17,7 +17,7 @@ import FormatListBulletedOutlinedIcon from '@mui/icons-material/FormatListBullet
 import UpdateOutlinedIcon from '@mui/icons-material/UpdateOutlined';
 import WhatshotSharpIcon from '@mui/icons-material/WhatshotSharp';
 
-import { drawerWidth } from 'src/layouts/Main/Main';
+import { drawerWidth } from 'src/layouts/Main';
 import { Profile } from 'src/components//Profile/Profile';
 
 type TNavTypes = {
